update: save balance info

+++ programs/saturn-v1/src/instructions/klend_lend.rs
@@ -119,7 +119,7 @@ pub fn handle(ctx: Context<KaminoLend>, amount: u64) -> Result<()> {
     )?;
 
     let treasury = &mut ctx.accounts.treasury;
-    treasury.kamino_lend_amount += amount as i64;
+    treasury.kamino_lend_amount += amount;
     treasury.treasury_value -= amount as u64;
 
     Ok(())
