+++ programs/saturn-v1/src/instructions/klend_withdraw.rs
@@ -114,7 +114,7 @@ pub fn handle(ctx: Context<KlendWithdraw>, amount: u64) -> Result<()> {
     )?;
 
     let treasury = &mut ctx.accounts.treasury;
-    treasury.kamino_lend_amount -= amount as i64;
+    treasury.kamino_lend_amount -= amount;
     treasury.treasury_value += amount as u64;
 
     Ok(())
