+++ programs/saturn-v1/src/instructions/marginfi_withdraw.rs
@@ -92,7 +92,7 @@ pub fn handle(ctx: Context<MarginfiWithdraw>, amount: u64) -> Result<()> {
     );
 
     let treasury = &mut ctx.accounts.treasury;
-    treasury.marginfi_lend_amount -= amount as i64;
+    treasury.marginfi_lend_amount -= amount;
     treasury.treasury_value += amount as u64;
 
     Ok(())
